+++ crates/vm/src/system/memory/manager/dimensions.rs
@@ -12,7 +12,7 @@ pub struct MemoryDimensions {
     /// Pointer height
     pub address_height: usize,
     /// Address space offset
-    pub as_offset: usize,
+    pub as_offset: u32,
 }
 
 impl MemoryDimensions {
@@ -20,9 +20,9 @@ impl MemoryDimensions {
         self.as_height + self.address_height
     }
     /// Convert an address label (address space, block id) to its index in the memory merkle tree.
-    pub fn label_to_index(&self, label: (usize, usize)) -> usize {
+    pub fn label_to_index(&self, label: (u32, u32)) -> u64 {
         let (addr_space, block_id) = label;
-        ((addr_space - self.as_offset) << self.address_height) + block_id
+        (((addr_space - self.as_offset) as u64) << self.address_height) + block_id as u64
     }
 }
