+++ crates/vm/src/system/memory/merkle/mod.rs
@@ -17,7 +17,7 @@ mod tests;
 #[derive(Debug)]
 pub struct MemoryMerkleChip<const CHUNK: usize, F> {
     pub air: MemoryMerkleAir<CHUNK>,
-    touched_nodes: FxHashSet<(usize, usize, usize)>,
+    touched_nodes: FxHashSet<(usize, u32, u32)>,
     num_touched_nonleaves: usize,
     final_state: Option<FinalState<CHUNK, F>>,
     overridden_height: Option<usize>,
@@ -56,7 +56,7 @@ impl<const CHUNK: usize, F: PrimeField32> MemoryMerkleChip<CHUNK, F> {
         self.overridden_height = Some(override_height);
     }
 
-    fn touch_node(&mut self, height: usize, as_label: usize, address_label: usize) {
+    fn touch_node(&mut self, height: usize, as_label: u32, address_label: u32) {
         if self.touched_nodes.insert((height, as_label, address_label)) {
             assert_ne!(height, self.air.memory_dimensions.overall_height());
             if height != 0 {
@@ -70,11 +70,11 @@ impl<const CHUNK: usize, F: PrimeField32> MemoryMerkleChip<CHUNK, F> {
         }
     }
 
-    pub fn touch_address(&mut self, address_space: usize, address: usize) {
+    pub fn touch_address(&mut self, address_space: u32, address: u32) {
         self.touch_node(
             0,
             address_space - self.air.memory_dimensions.as_offset,
-            address / CHUNK,
+            address / CHUNK as u32,
         );
     }
 }
